Assignment 5/eb-flask/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
mysql-connector
numpy
scikit-learn